refactor: LLM 유틸 구조 정비 및 서비스 역할 분리 개선
- get_summarize_response: 프롬프트 기반 호출로 명확히 리네이밍
- get_chat_response: 프롬프트 인자로 받도록 수정하여 유연성 확보
- build_prompt, llm_generate 함수 호출 흐름 단순화
- 파일명 vector_store.py → vector_loader.py로 변경
