+++ app/models/llm_client.py
@@ -1,6 +1,6 @@
 from app.models.qwen2_5_loader import tokenizer, llm, sampling_params
 from app.models.prompt_template import chatbot_rag_prompt
-from app.core.vector_store import load_vectorstore
+from app.core.vector_loader import load_vectorstore
 from dotenv import load_dotenv
 from langsmith import traceable
 from langsmith.run_helpers import get_current_run_tree
@@ -10,10 +10,6 @@ load_dotenv()
 
 # 시스템 메시지
 SYSTEM_MESSAGE = "You are a kind and friendly chatbot for announcements. Answer politely and clearly based on the announcement content. Please respond only in Korean."
-
-# 벡터스토어 로딩 후 RAG용 retriever 구성
-faiss_vectorstore = load_vectorstore()
-retriever = faiss_vectorstore.as_retriever(search_kwargs={"k": 2})
 
 # 공통 메시지 생성 및 프롬프트 구성 함수
 def build_prompt(user_input: str, context: str = "") -> str:
@@ -45,19 +41,15 @@ async def llm_generate(prompt_str: str, request_id: str) -> str:
     result = await get_last_output(agen)
     return result if result else "empty:" + prompt_str
 
-# 요약/뉴스 생성 전용 호출 함수
-@traceable(name="요약/뉴스 생성 호출", inputs={"프롬프트": lambda args, kwargs: args[0]})
-async def call_qwen(prompt: str, request_id: str) -> str:
+# 요약/뉴스 생성 전용 함수
+@traceable(name="요약 및 헤드라인 응답", inputs={"프롬프트": lambda args, kwargs: args[0]})
+async def get_summarize_response(prompt: str, request_id: str) -> str:
     prompt_str = build_prompt(prompt)
     return await llm_generate(prompt_str, request_id)
 
 # 문서 기반 챗봇 응답 함수
 @traceable(name="챗봇 질문 응답", inputs={"질문": lambda args, kwargs: args[0]})
-async def get_chat_response(question: str, request_id: str) -> str:
-    parsed_question = parse_relative_dates(question)
-    docs = retriever.get_relevant_documents(question)
-    context = "\n\n".join([doc.page_content for doc in docs])
-    prompt = chatbot_rag_prompt.format(context=context, question=parsed_question)
+async def get_chat_response(prompt: str, docs: str, request_id: str) -> str:
     prompt_str = build_prompt(prompt)
 
     run = get_current_run_tree()
