+++ app/services/notice_service.py
@@ -1,6 +1,6 @@
 import json
-from app.core.llm_client import call_qwen
-from app.model.prompt_template import notice_summary_prompt
+from app.models.llm_client import get_summarize_response
+from app.models.prompt_template import notice_summary_prompt
 
 async def summarize_notice_service(title: str, content: str, request_id: str) -> str:
     # 공지 요약 프롬프트 구성 (title은 포함하지 않고 content만 사용)
@@ -8,7 +8,7 @@ async def summarize_notice_service(title: str, content: str, request_id: str) ->
 
     isCompleted = True
     # LLM 응답 호출
-    response = await call_qwen(prompt, request_id)
+    response = await get_summarize_response(prompt, request_id)
 
     # JSON 파싱
     try:
